+++ src/fe_task/serializers.py
@@ -7,4 +7,4 @@ class TaskSerializer(serializers.ModelSerializer):
     class Meta:
         model = Task
         read_only_fields = ('uuid',)
-        fields = ('uuid', 'created_at', 'updated_at', 'name', 'description')
+        fields = ('uuid', 'created_at', 'updated_at', 'name', 'parent', 'description',)