+++ src/fe_task/tests/test_task.py
@@ -7,3 +7,11 @@ class TestTask(TestCase):
     def test_factory(self):
         task = TaskFactory()
         assert task is not None
+
+    def test_field_parent(self):
+        task = TaskFactory()
+        assert hasattr(task, 'parent')
+
+    def test__str__(self):
+        task = TaskFactory(name='test__str__')
+        assert 'test__str__' == str(task)