+++ src/fe_task/tests/test_taskModelViewSet.py
@@ -1,3 +1,5 @@
+import uuid
+
 from django.urls import reverse
 from fe_core.factories import UserFactory, EntityFactory
 from rest_framework import status
@@ -29,14 +31,22 @@ class TestTaskModelViewSet(APITestCase):
         assert status.HTTP_201_CREATED == response.status_code
         task = Task.objects.first()
         data = response.data
-        assert 5 == len(data)
+        assert 6 == len(data)
         assert str(task.uuid) == data['uuid']
         assert task.name == data['name']
         assert task.description == data['description']
         assert task.entity == self.entity
         assert task.user == self.user
+        assert task.parent is None
 
     def test_anonymous_user(self):
         self.client.logout()
         response = self.client.post(reverse('task-list'), EXAMPLE_DATA)
         assert status.HTTP_403_FORBIDDEN == response.status_code
+
+    def test_create_field_with_parent(self):
+        data = EXAMPLE_DATA.copy()
+        parent_uuid = str(uuid.uuid4())
+        data['parent'] = parent_uuid
+        response = self.client.post(reverse('task-list'), data)
+        assert response.data['parent'] == parent_uuid